[!!!][FEATURE] Breaking change! Remove "mode" setting and introduce possibility to set all configuration by "exceptions" only. Make backend user authorization as regular condition.
