+++ Configuration/TCA/Overrides/be_users.php
@@ -3,7 +3,7 @@
 $tempColumns = [
     'tx_restrictfe_clearbesession' => [
         'exclude' => 1,
-        'label' => 'Clear BE session after login.',
+        'label' => 'Clear BE session after login',
         'config' => [
             'type' => 'check',
             'default' => 0
