+++ Resources/Private/Templates/Restricted.html
@@ -3,7 +3,7 @@
         "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
 <html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
 <head>
-    <title>Bitte melden Sie sich an um diese Seite sehen zu können.</title>
+    <title>Please login first to see this website / Bitte melden Sie sich an um diese Seite sehen zu können.</title>
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8"/>
     <meta name="robots" content="noindex, nofollow">
 
@@ -11,7 +11,8 @@
 <body style="text-align: center">
 <p>
     <br/><br/><br/>
-    Bitte <a href="{beLoginLink}">melden Sie sich an um</a> diese Seite sehen zu können.
+    [EN] Please <a href="{beLoginLink}">login</a> first to see this website.<br>
+    [DE] Bitte <a href="{beLoginLink}">melden Sie sich an um</a> diese Seite sehen zu können.
 </p>
 </body>
 </html>
