+++ ext_localconf.php
@@ -1,6 +1,7 @@
 <?php
 
 if (TYPO3_MODE == 'FE') {
-    $TYPO3_CONF_VARS['SC_OPTIONS']['tslib/class.tslib_fe.php']['checkAlternativeIdMethods-PostProc'][]
-        = 'SourceBroker\Restrictfe\Main->redirectCheckForLoggedBeUser';
+    // hook into settingLanguage_postProcess to have info about sys_language_uid to have ability to do conditions with language
+    $TYPO3_CONF_VARS['SC_OPTIONS']['tslib/class.tslib_fe.php']['settingLanguage_postProcess'][]
+        = 'SourceBroker\Restrictfe\Restrict->restrictFrontend';
 }
